Fix bug caused by sign extension when reading single-byte row numbers; fix some bugs in new text index generation code
git-svn-id: https://svn.code.sf.net/p/jackcess/code/jackcess/trunk@252 f203690c-595d-4dc9-a70b-905162fa7fd2

+++ src/java/com/healthmarketscience/jackcess/UsageMap.java
@@ -115,7 +115,7 @@ public class UsageMap
    *         which type of map is found
    */
   public static UsageMap read(Database database, int pageNum,
-                              byte rowNum, boolean assumeOutOfRangeBitsOn)
+                              int rowNum, boolean assumeOutOfRangeBitsOn)
     throws IOException
   {
     JetFormat format = database.getFormat();
